Working app state transfers. Improved socket connection/disconnection. Made messages relational to users.
