+++ src/lib/components/forms/JoinRoomForm.svelte
@@ -1,61 +1,9 @@
 <script lang="ts">
 	import { socket } from '$lib/sockets/client';
-	import { goto } from '$app/navigation';
-	import { members } from '$lib/stores/members-store';
 	import { joiningRoom } from '$lib/stores/socket-store';
-	import { user } from '$lib/stores/user-store';
-	import { onMount } from 'svelte';
 	import { uniqueNamesGenerator, adjectives, animals, type Config } from 'unique-names-generator';
-	import type {
-		InvalidDataEvent,
-		JoinRoomEvent,
-		RoomJoinedEvent,
-		RoomNotFoundEvent
-	} from '$lib/types/socket-types';
+	import type { JoinRoomEvent } from '$lib/types/socket-types';
 	import { validJoinRoom } from '$lib/validation/processors';
-
-	function handleErrorMessage(message: string) {
-		console.error(message);
-	}
-
-	function onRoomJoined(event: RoomJoinedEvent) {
-		console.debug(event);
-		$user = event?.user;
-		$members = event?.members;
-		goto(`/${event?.roomId}`);
-		$joiningRoom = false;
-	}
-
-	function onRoomNotFound(event: RoomNotFoundEvent) {
-		console.debug(event);
-		if ($joiningRoom) {
-			$joiningRoom = false;
-			handleErrorMessage(event?.message);
-		}
-	}
-
-	function onInvalidData(event: InvalidDataEvent) {
-		console.debug(event);
-		if ($joiningRoom) {
-			$joiningRoom = false;
-			handleErrorMessage(event?.message);
-		}
-	}
-
-	onMount(() => {
-		socket.on('room-joined', onRoomJoined);
-		socket.on('room-not-found', onRoomNotFound);
-		socket.on('invalid-data', onInvalidData);
-
-		return () => {
-			console.debug(`socket-off:room-joined:${socket.id}`);
-			console.debug(`socket-off:room-not-found:${socket.id}`);
-			console.debug(`socket-off:invalid-data:${socket.id}`);
-			socket.off('room-joined', onRoomJoined);
-			socket.off('room-not-found', onRoomNotFound);
-			socket.off('invalid-data', handleErrorMessage);
-		};
-	});
 
 	function submit() {
 		$joiningRoom = true;
